Renames package utils -> util
